Move dirTree filter functions to separate module
