+++ index.js
@@ -7,6 +7,8 @@ const dirTreeToArchyTree = require('./dirTreeToArchyTree');
 const wrapHtml = require('./wrapHtml');
 const { DirectoryInvalidError } = require('./exceptions');
 
+const { filterToMaxDepth, filterIncluded, filterEmptyDirectories } = require('./dirTreeFilters')
+
 const defaultOpts = {
     fileTypes: null,
     isHtml: false,
@@ -18,59 +20,7 @@ const defaultOpts = {
 };
 
 /**
- * Remove all tree nodes below a given depth.
- * Note: This feature can be inefficient due to this filtering taking place on a complete directory tree,
- * compared to filtering while building the dirTree in the first place.
- * To fix we would need to add this feature to directory-tree instead.
- * @param {dirTree} tree A directory tree
- * @param {number} maxDepth Maximum depth of files/directories to include in tree
- */
-function filterToMaxDepth(tree, maxDepth) {
-    if (tree.children && tree.children.length > 0) {
-        if (maxDepth <= 0) {
-            tree.children = []
-        } else {
-            tree.children.forEach(child => {
-                if (child.type === 'directory') {
-                    filterToMaxDepth(child, maxDepth - 1);
-                }
-            })
-        }
-    }
-}
-
-/**
- * Remove files and directories from tree that don't match regexp.
- * Note: This feature can be inefficient due to this filtering taking place on a complete directory tree,
- * compared to filtering while building the dirTree in the first place.
- * To fix we would need to add this feature to directory-tree instead.
- * @param {dirTree} tree A directory tree
- * @param {Regexp} regexp Regexp to match nodes against
- */
-function filterIncluded(tree, regexp) {
-    if (!tree || !regexp.test(tree.name)) {
-        return false;
-    }
-    if (tree.children && tree.children.length > 0) {
-        tree.children = tree.children.filter(child => filterIncluded(child, regexp));
-    }
-    return true;
-}
-
-function filterEmptyDirectories(tree) {
-    if (tree.children && tree.children.length > 0) {
-        tree.children.forEach(child => {
-            if (child.type === 'directory') {
-                child.children = filterEmptyDirectories(child);
-            }
-        });
-    }
-    return tree.children = tree.children.filter(child => child.type === 'file' || child.children.length > 0);
-}
-
-/**
 * Generates a directory tree from the given directory and all sub-directories
-*
 * @param {!String} dir The directory to use as the start (this will be the root node of the tree)
 * @param {Object} [opts] An object which supports the following options:
 *                        {Array.<String>} fileTypes The file types to print. Defaults to all file types.
